fix : 게시글 수정 버그 수정

+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/dto/response/ResponseGetPostList.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record ResponseGetPostList(
         Long memberId,
         String title,
-        Long hits,
+        Integer hits,
         LocalDateTime createdAt
-   ) {
+) {
 }
